Split shell computer observe tests
